Rename io.py to filemanip.py
This renaming aims to solve Issue #37. io.py name can generate conflict with IDE.
