+++ clinica/utils/filemanip.py
